transitioned angular-xeditable
