+++ angular-xeditable/angular-xeditable-tests.ts
@@ -1,9 +1,9 @@
-﻿/// <reference path="angular-xeditable.d.ts" />
+﻿/// <reference types="angular-xeditable" />
 
 var myApp = angular.module('testModule', ['xeditable']);
 
 myApp.run(["editableOptions", (editableOptions: angular.xeditable.IEditableOptions) => {
-    
+
         editableOptions.activate = "select";
         editableOptions.activationEvent = "click";
         editableOptions.blurElem = "ignore";
